requirements.txt
# Gianluigi Dependencies
# Python 3.11+ required

# =============================================================================
# Core AI & LLM
# =============================================================================
anthropic>=0.39.0              # Claude API client
openai>=1.50.0                 # For embeddings (text-embedding-3-small)

# =============================================================================
# Database
# =============================================================================
supabase>=2.9.0                # Supabase client (PostgreSQL + pgvector)

# =============================================================================
# Google APIs
# =============================================================================
google-api-python-client>=2.150.0   # Google API client library
google-auth-httplib2>=0.2.0         # Auth for Google APIs
google-auth-oauthlib>=1.2.0         # OAuth for Google APIs

# =============================================================================
# Telegram
# =============================================================================
python-telegram-bot>=21.7      # Telegram Bot API

# =============================================================================
# Data Validation & Models
# =============================================================================
pydantic>=2.9.0                # Data validation and settings
pydantic-settings>=2.5.0       # Settings management with Pydantic

# =============================================================================
# Environment & Configuration
# =============================================================================
python-dotenv>=1.0.0           # Load .env files

# =============================================================================
# HTTP & Async
# =============================================================================
httpx>=0.27.0                  # Async HTTP client
aiohttp>=3.9.0                 # HTTP server for health checks (Cloud Run)
aiofiles>=24.1.0               # Async file operations
mcp>=1.26.0                    # MCP Python SDK (Streamable HTTP transport for Claude.ai)
uvicorn>=0.30.0                # ASGI server for MCP Starlette app

# =============================================================================
# Document Processing
# =============================================================================
PyPDF2>=3.0.0                  # PDF text extraction
python-docx>=1.1.0             # Word document processing
python-pptx>=1.0.0             # PowerPoint slide extraction
jinja2>=3.1.0                  # HTML report templating (v1.0)

# =============================================================================
# Utilities
# =============================================================================
tenacity>=9.0.0                # Retry logic for API calls
python-dateutil>=2.9.0         # Date parsing utilities

# =============================================================================
# Development & Testing
# =============================================================================
pytest>=8.3.0                  # Testing framework
pytest-asyncio>=0.24.0         # Async test support
pytest-cov>=4.1.0              # Coverage reporting
# black>=24.8.0                # Code formatting (optional)
# ruff>=0.6.0                  # Linting (optional)
# mypy>=1.11.0                 # Type checking (optional)

# Intelligence Signal — video pipeline
Pillow>=10.4.0                   # Slide image generation for video
